+++ lib/output/constructor.go
@@ -70,6 +70,7 @@ const (
 	TypeHTTPServer    = "http_server"
 	TypeInproc        = "inproc"
 	TypeKafka         = "kafka"
+	TypeKinesis       = "kinesis"
 	TypeMQTT          = "mqtt"
 	TypeNanomsg       = "nanomsg"
 	TypeNATS          = "nats"
@@ -77,6 +78,7 @@ const (
 	TypeNSQ           = "nsq"
 	TypeRedisList     = "redis_list"
 	TypeRedisPubSub   = "redis_pubsub"
+	TypeRedisStreams  = "redis_streams"
 	TypeS3            = "s3"
 	TypeSQS           = "sqs"
 	TypeSTDOUT        = "stdout"
@@ -99,13 +101,15 @@ type Config struct {
 	HTTPServer    HTTPServerConfig           `json:"http_server" yaml:"http_server"`
 	Inproc        InprocConfig               `json:"inproc" yaml:"inproc"`
 	Kafka         writer.KafkaConfig         `json:"kafka" yaml:"kafka"`
+	Kinesis       writer.KinesisConfig       `json:"kinesis" yaml:"kinesis"`
 	MQTT          writer.MQTTConfig          `json:"mqtt" yaml:"mqtt"`
 	Nanomsg       NanomsgConfig              `json:"nanomsg" yaml:"nanomsg"`
 	NATS          NATSConfig                 `json:"nats" yaml:"nats"`
 	NATSStream    NATSStreamConfig           `json:"nats_stream" yaml:"nats_stream"`
 	NSQ           NSQConfig                  `json:"nsq" yaml:"nsq"`
 	RedisList     writer.RedisListConfig     `json:"redis_list" yaml:"redis_list"`
-	RedisPubSub   RedisPubSubConfig          `json:"redis_pubsub" yaml:"redis_pubsub"`
+	RedisPubSub   writer.RedisPubSubConfig   `json:"redis_pubsub" yaml:"redis_pubsub"`
+	RedisStreams  writer.RedisStreamsConfig  `json:"redis_streams" yaml:"redis_streams"`
 	S3            writer.AmazonS3Config      `json:"s3" yaml:"s3"`
 	SQS           writer.AmazonSQSConfig     `json:"sqs" yaml:"sqs"`
 	STDOUT        STDOUTConfig               `json:"stdout" yaml:"stdout"`
@@ -128,13 +132,15 @@ func NewConfig() Config {
 		HTTPServer:    NewHTTPServerConfig(),
 		Inproc:        NewInprocConfig(),
 		Kafka:         writer.NewKafkaConfig(),
+		Kinesis:       writer.NewKinesisConfig(),
 		MQTT:          writer.NewMQTTConfig(),
 		Nanomsg:       NewNanomsgConfig(),
 		NATS:          NewNATSConfig(),
 		NATSStream:    NewNATSStreamConfig(),
 		NSQ:           NewNSQConfig(),
 		RedisList:     writer.NewRedisListConfig(),
-		RedisPubSub:   NewRedisPubSubConfig(),
+		RedisPubSub:   writer.NewRedisPubSubConfig(),
+		RedisStreams:  writer.NewRedisStreamsConfig(),
 		S3:            writer.NewAmazonS3Config(),
 		SQS:           writer.NewAmazonSQSConfig(),
 		STDOUT:        NewSTDOUTConfig(),